Move the directory names
