+++ run_benchmark.py
@@ -26,7 +26,7 @@ def run():
     args, bm_args = parser.parse_known_args()
 
     try:
-        benchmark = importlib.import_module(f"userbenchmark.{available_benchmarks[args.bm_name]}")
+        benchmark = importlib.import_module(f"userbenchmark.{available_benchmarks[args.bm_name]}.run")
         benchmark.run(bm_args)
     except ImportError as e:
         print(f"Failed to import user benchmark module {args.bm_name}, error: {str(e)}")
